+++ README.md
@@ -71,8 +71,40 @@
     search,
     page,
     page_size,
-    short,
-    order_by.
+    short.
+
+    Examples:
+
+    Get first page of authors:
+        GET
+        http://127.0.0.1:8000/song/author/
+
+    Get first page of authors with just some (reduced) data fields count:
+        GET
+        http://127.0.0.1:8000/song/author/?short=1
+
+    Get authors count with 'Mr' in name:
+        GET
+        http://127.0.0.1:8000/song/author/?search=Mr&page=0
+
+    Get authors page 10:
+        GET
+        http://127.0.0.1:8000/song/author/?page=10&page_size=5
+
+
+    Create an author 'new author':
+        POST body={'name'='new author'}
+        http://127.0.0.1:8000/song/author/
+
+
+    Update the name of the author with id=1 to 'new name:
+        PUT body={'name'='new name'}
+        http://127.0.0.1:8000/song/author/1/
+
+
+    Delete author with id=1:
+        DELETE
+        http://127.0.0.1:8000/song/author/1/
 
 <p>
     Schemas:
